requirements.txt
asgiref==3.2.10
autopep8==1.5.4
dj-database-url==0.5.0
Django==2.2
djangorestframework==3.12.1
gunicorn==20.0.4
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
